refactor: separate helpers from use-merged-widget-state
