+++ apps/web/src/services/dashboards/widgets/_hooks/use-widget/use-merged-widget-state.ts
@@ -3,19 +3,19 @@ import {
     computed, reactive,
 } from 'vue';
 
-import { merge } from 'lodash';
-
 import type { DashboardSettings, DashboardVariables, DashboardVariablesSchema } from '@/services/dashboards/config';
 import type {
     WidgetConfig, WidgetOptions,
     InheritOptions,
 } from '@/services/dashboards/widgets/_configs/config';
-import { getWidgetFilterDataKey } from '@/services/dashboards/widgets/_helpers/widget-filters-helper';
 import { getWidgetConfig } from '@/services/dashboards/widgets/_helpers/widget-helper';
+import { getMergedWidgetInheritOptions } from '@/services/dashboards/widgets/_helpers/widget-inherit-options-helper';
+import {
+    getRefinedWidgetOptions,
+} from '@/services/dashboards/widgets/_helpers/widget-options-helper';
 import {
     getInitialSchemaProperties,
 } from '@/services/dashboards/widgets/_helpers/widget-schema-helper';
-import type { InheritOptionsErrorMap } from '@/services/dashboards/widgets/_helpers/widget-validation-helper';
 import { getWidgetInheritOptionsErrorMap } from '@/services/dashboards/widgets/_helpers/widget-validation-helper';
 
 
@@ -89,61 +89,7 @@ export function useMergedWidgetState(
 }
 
 
-const getRefinedWidgetOptions = (
-    widgetConfig?: WidgetConfig,
-    optionsData?: WidgetOptions,
-    mergedInheritOptions?: InheritOptions,
-    dashboardVariables?: DashboardVariables,
-    optionsErrorMap?: InheritOptionsErrorMap,
-): WidgetOptions => {
-    const mergedOptions = getMergedWidgetOptions(widgetConfig, optionsData);
-    if (!mergedInheritOptions || !dashboardVariables) return mergedOptions;
 
-    const parentOptions: Partial<WidgetOptions> = getRefinedParentOptions(mergedInheritOptions, dashboardVariables, optionsErrorMap);
-    const refined = merge({}, mergedOptions, parentOptions);
-    return refined;
-};
 
-const getMergedWidgetOptions = (widgetConfig?: WidgetConfig, widgetOptions?: WidgetOptions) => merge({}, widgetConfig?.options ?? {}, widgetOptions);
 
-const getMergedWidgetInheritOptions = (widgetConfig?: WidgetConfig, inheritOptionsData?: InheritOptions): InheritOptions => {
-    const mergedInheritOptions = merge({}, widgetConfig?.inherit_options ?? {}, inheritOptionsData);
-    Object.values(mergedInheritOptions).forEach((inheritOption) => {
-        if (inheritOption.enabled === false && inheritOption.variable_info) {
-            inheritOption.variable_info = undefined;
-        }
-    });
-    return mergedInheritOptions;
-};
-
-const getRefinedParentOptions = (
-    inheritOptions: InheritOptions,
-    dashboardVariables: DashboardVariables,
-    optionsErrorMap?: InheritOptionsErrorMap,
-): Partial<WidgetOptions> => {
-    const result: Partial<WidgetOptions> = {
-        filters: {},
-    };
-    Object.entries(inheritOptions).forEach(([filterKey, inheritOption]) => {
-        if (optionsErrorMap?.[filterKey]) return;
-
-        const variableKey = inheritOption?.variable_info?.key;
-        if (!inheritOption?.enabled || !variableKey) return;
-
-        const variableValue = dashboardVariables[variableKey];
-        if (!variableValue || !variableValue?.length) return;
-
-        if (filterKey.startsWith('filters.')) {
-            const _filterKey = filterKey.replace('filters.', '');
-            const filterDataKey = getWidgetFilterDataKey(_filterKey);
-            result.filters = {
-                ...result.filters,
-                [_filterKey]: [{ k: filterDataKey, v: variableValue, o: '=' }],
-            };
-        } else {
-            result[filterKey] = variableValue;
-        }
-    });
-    return result;
-};
